Update README.md
New PR because of 
-Conflicts and 
-a failure in the version number and 
I was not able to change the PR: https://github.com/joomla-projects/selenium-server-standalone/pull/17

I changed the number of the version in the prior PR to "v2.53.1" because I find the line 
	"joomla-projects/selenium-server-standalone": "v2.53.1",
in composer.json in the folder /tests/codeception. 

Is the number v2.42 or v2.53.1 correct?