Update README.md
Today I installed everything new. 
We do not need the changes in robofile.php any more. And the global installation is not needed (I am not sure if I was wrong when I added this suggestion here. But today it worked for me without the global installation).
I do not know where I can check the version number. But because of the file size I am sure that 2.25 (and not the current 2.26) is OK.